samples: address Copilot PR #8 review comments
- CMakeLists.txt (all samples + native_sim_harness): replace bare
  EXTRA_ZEPHYR_MODULES append with an EXISTS-guarded block that
  falls back to the repo root module when the local modules/zephyrdb-local
  symlink is absent.
- ts_basic/src/main.c: print aggregate value with %%f (not %%d) since
  zdb_ts_agg_result_t::value is a double.
- doc_query_filters/src/main.c: call zdb_doc_metadata_free() after a
  successful query to release allocated field-name strings.
- kv_basic/src/main.c: normalise indentation to tabs throughout main()
  and the cfg initialiser for consistency with project style.

+++ samples/ts_basic/src/main.c
@@ -61,7 +61,7 @@ int main(void)
 
     rc = zdb_ts_query_aggregate(&stream, window, ZDB_TS_AGG_AVG, &agg);
     if (rc == ZDB_OK) {
-        printk("TS sample PASS: avg=%d points=%u\n", (int)agg.value, agg.points);
+        printk("TS sample PASS: avg=%f points=%u\n", agg.value, agg.points);
     } else {
         printk("TS sample: aggregate query returned rc=%d\n", (int)rc);
     }
